research/endpoint_tests/requirements.txt
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
polars>=0.19.0
pyarrow>=14.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
notebook>=7.0.0
